Potential fix for code scanning alert no. 6: Implicit narrowing conversion in compound assignment

+++ tools/dmitool/src/main/java/dmitool/DMI.java
@@ -27,7 +27,7 @@ import java.util.List;
 public class DMI implements Comparator<IconState> {
     int w, h;
     List<IconState> images;
-    int totalImages = 0;
+    long totalImages = 0;
     RGBA[] palette;
     boolean isPaletted;
